fix camel casing module names
